chore(nip-tags): create separate module for nip tags
