+++ Dockerfile
@@ -2,5 +2,5 @@
 # Base image
 FROM postgres:17-bookworm
 
-# Copy the sql script we wann use over
-COPY ./init/init_nips_tags.sql ./docker-entrypoint-initdb.d
+# Copy the sql script we want use over
+COPY ./nostr-protocol-nips_nip-tags/init/init.sql ./docker-entrypoint-initdb.d
