+++ nostr-protocol-nips_nip-tags/pkg/nipTags/nip-tags.go
@@ -1,4 +1,4 @@
-package scripts
+package nipTags
 
 import (
 	"fmt"
